+++ src/function/flip.js
@@ -20,4 +20,28 @@ export const
      * @param fn {Function}
      * @returns {Function}
      */
-    flip = fn => curry((b, a) => call(fn, a, b));
+    flip = fn => curry((b, a) => call(fn, a, b)),
+
+    /**
+     * Same as `flip` except returns a flipped function of arity 3.
+     * @function module:function.flip3
+     * @param fn {Function}
+     * @returns {Function}
+     */
+    flip3 = fn => curry((c, b, a) => call(fn, a, b, c)),
+
+    /**
+     * Same as `flip` except returns a flipped function of arity 4.
+     * @function module:function.flip4
+     * @param fn {Function}
+     * @returns {Function}
+     */
+    flip4 = fn => curry((d, c, b, a) => call(fn, a, b, c, d)),
+
+    /**
+     * Same as `flip` except returns a flipped function of arity 5.
+     * @function module:function.flip5
+     * @param fn {Function}
+     * @returns {Function}
+     */
+    flip5 = fn => curry((e, d, c, b, a) => call(fn, a, b, c, d, e));