+++ types/utils.ts
@@ -1,11 +1,11 @@
-declare function fPureTakesOne(arg: any, obj: any): any;
+declare function fPureTakesOne(name): (arg: any, obj: any) => any;
 
-declare function fPureTakes2(arg1: any, arg2: any, obj: any): any;
+declare function fPureTakes2(name): (arg1: any, arg2: any, obj: any) => any;
 
-declare function fPureTakes3(arg1: any, arg2: any, arg3: any, obj: any): any;
+declare function fPureTakes3(name): (arg1: any, arg2: any, arg3: any, obj: any) => any;
 
-declare function fPureTakes4(arg1: any, arg2: any, arg3: any, arg4: any, obj: any): any;
+declare function fPureTakes4(name): (arg1: any, arg2: any, arg3: any, arg4: any, obj: any) => any;
 
-declare function fPureTakes5(arg1: any, arg2: any, arg3: any, arg4: any, arg5: any, obj: any): any;
+declare function fPureTakes5(name): (arg1: any, arg2: any, arg3: any, arg4: any, arg5: any, obj: any) => any;
 
-declare function fPureTakesOneOrMore(obj: any, ...args: any[]): any;
+declare function fPureTakesOneOrMore(name): (obj: any, ...args: any[]) => any;